Remove style and copy contexts
This feels overwhelming, just pass as props
