+++ packages/notifi-react-card/lib/components/NotifiFooter.tsx
@@ -1,19 +1,30 @@
-import { useNotifiCopyContext, useNotifiStyleContext } from '../context';
+import type { DeepPartialReadonly } from '../utils';
 import { NotifiLogo } from './NotifiLogo';
 import React from 'react';
 
-export const NotifiFooter: React.FC = () => {
-  const { footer: copy } = useNotifiCopyContext();
-  const { footer: styles } = useNotifiStyleContext();
+type Props = Readonly<{
+  classNames?: DeepPartialReadonly<{
+    container: string;
+    poweredBy: string;
+    logoSvg: string;
+    link: string;
+    spacer: string;
+  }>;
+  copy?: DeepPartialReadonly<{
+    poweredBy: string;
+    learnMore: string;
+  }>;
+}>;
 
+export const NotifiFooter: React.FC<Props> = ({ classNames, copy }: Props) => {
   return (
-    <div className={styles?.container}>
-      <span className={styles?.poweredBy}>
+    <div className={classNames?.container}>
+      <span className={classNames?.poweredBy}>
         {copy?.poweredBy ?? 'Powered by'}
       </span>
-      <NotifiLogo className={styles?.logoSvg} />
-      <span className={styles?.spacer} />
-      <span className={styles?.link}>
+      <NotifiLogo className={classNames?.logoSvg} />
+      <span className={classNames?.spacer} />
+      <span className={classNames?.link}>
         <a
           href="https://docs.notifi.network/NotifiIntegrationsFAQ.html"
           target="_blank"
